pyproject.toml
[tool.poetry]
name = "alphabetsoup"
version = "0.5.4-post.1+472170d"
description = "tile phylogenetic space with subtrees"
authors = ["Joel Berendzen <joelb@ncgr.org>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/ncgr/alphabetsoup"
documentation = "https://alphabetsoup.readthedocs.org"
keywords = [
    "science",
    "biology",
    "bioinformatics",
    "genomics",
    "data science",
    "phylogeny",
    "protein",
    "sequences",
    "FASTA"
    ]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Console',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: BSD License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Topic :: Scientific/Engineering :: Bio-Informatics'
    ]

[tool.poetry.dependencies]
python = "^3.6.1"
biopython = "^1.76"
click = "^7.1.2"
dask = {extras = ["bag"], version = "^2.15.0"}
matplotlib = "^3.2.1"
pandas = "^1.0.3"
seaborn = "^0.10.1"
click_loguru = "^0.3.3"
toolz = "^0.10.0"
cloudpickle = "^1.4.1"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.8.1"
black = {version = "^19.10b0", allow-prereleases = true}
pre-commit = "^2.3.0"
pydocstyle = "^5.0.2"
isort = {version = "^4.3.21", extras = ["pyproject"]}
sh = "^1.13.0"
pylint = "^2.5.2"
coverage = {extras = ["toml"], version = "^5.1"}
pytest-datadir-mgr = "^1.0.1"
nox = "^2020.5.24"
flynt = "^0.48.1"

[tool.poetry.scripts]
alphabetsoup = "alphabetsoup:cli"

[tool.black]
line-length = 79
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
  \.git
  | \.mypy
  _cache
  | \.nox
  | \.idea
  | \.pytest-cache
  | _build
  | build
  | dist
  | tests/
)/
'''

[tool.isort]
atomic = true
force_single_line = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
import_heading_stdlib = "standard library imports"
import_heading_thirdparty = "third-party imports"
import_heading_firstparty = "first-party imports"
import_heading_localfolder = "module imports"
known_standard_library = "pkg_resources"

[tool.poetry-dynamic-versioning]
enable = false
style = "semver"

[tool.coverage.run]
parallel = true
omit = ["tests/*", "noxfile.py", ".nox/*"]

[tool.coverage.report]
exclude_lines = [
    "if self.debug:",
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
    ]


[tool.pylint.'MESSAGES CONTROL']
disable = """,
   print-statement,
   parameter-unpacking,
   unpacking-in-except,
   old-raise-syntax,
   backtick,
   long-suffix,
   old-ne-operator,
   old-octal-literal,
   import-star-module-level,
   non-ascii-bytes-literal,
   raw-checker-failed,
   bad-inline-option,
   locally-disabled,
   file-ignored,
   suppressed-message,
   useless-suppression,
   deprecated-pragma,
   use-symbolic-message-instead,
   apply-builtin,
   basestring-builtin,
   buffer-builtin,
   cmp-builtin,
   coerce-builtin,
   execfile-builtin,
   file-builtin,
   long-builtin,
   raw_input-builtin,
   reduce-builtin,
   standarderror-builtin,
   unicode-builtin,
   xrange-builtin,
   coerce-method,
   delslice-method,
   getslice-method,
   setslice-method,
   no-absolute-import,
   old-division,
   dict-iter-method,
   dict-view-method,
   next-method-called,
   metaclass-assignment,
   indexing-exception,
   raising-string,
   reload-builtin,
   oct-method,
   hex-method,
   nonzero-method,
   cmp-method,
   input-builtin,
   round-builtin,
   intern-builtin,
   unichr-builtin,
   map-builtin-not-iterating,
   zip-builtin-not-iterating,
   range-builtin-not-iterating,
   filter-builtin-not-iterating,
   using-cmp-argument,
   eq-without-hash,
   div-method,
   idiv-method,
   rdiv-method,
   exception-message-attribute,
   invalid-str-codec,
   sys-max-int,
   bad-python3-import,
   deprecated-string-function,
   deprecated-str-translate-call,
   deprecated-itertools-function,
   deprecated-types-field,
   next-method-defined,
   dict-items-not-iterating,
   dict-keys-not-iterating,
   dict-values-not-iterating,
   deprecated-operator-function,
   deprecated-urllib-function,
   xreadlines-attribute,
   deprecated-sys-function,
   exception-escape,
   comprehension-escape,
   no-self-use,
   wrong-import-position,
   bad-continuation,
   too-many-locals,
   too-many-statements,
   too-few-public-methods,
   too-many-arguments,
   too-many-branches,
   too-many-instance-attributes,
   """

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
